init instance (#742)
* spec align report instance package 
* instance align
* proposal align
* prepare align
* commit align
* round change align

+++ scripts/spec_align_report/utils/file_system.go
@@ -1,10 +1,9 @@
 package utils
 
 import (
+	"github.com/pkg/errors"
 	"io"
 	"os"
-
-	"github.com/pkg/errors"
 )
 
 // Copy copies the contents of the file at srcpath to a regular file at dstpath.
@@ -15,7 +14,6 @@ func Copy(srcPath, dstPath string) (err error) {
 	if err != nil {
 		return err
 	}
-	//nolint: errcheck
 	defer r.Close() // ok to ignore error: file was opened read-only.
 
 	w, err := os.Create(dstPath)
@@ -39,8 +37,11 @@ func Copy(srcPath, dstPath string) (err error) {
 	return nil
 }
 
-func Mkdir(path string) (err error) {
-	os.RemoveAll(path)
+func Mkdir(path string, clean bool) (err error) {
+	if clean == true {
+		os.RemoveAll(path)
+	}
+
 	if _, err := os.Stat(path); errors.Is(err, os.ErrNotExist) {
 		err := os.Mkdir(path, os.ModePerm)
 		if err != nil {
